+++ src/run_preflight/__init__.py
@@ -2,6 +2,7 @@
 
 from .db import (
     AmpliconBarcodeRosterEntry,
+    AmpliconSampleRow,
     IlluminaSampleRow,
     KatharoseqSampleInfo,
     PacbioSampleRow,
@@ -9,6 +10,7 @@ from .db import (
     add_katharoseq_sample,
     create_db,
     get_amplicon_barcode_roster,
+    get_amplicon_sample_info,
     get_illumina_sample_info,
     get_katharoseq_sample_info,
     get_pacbio_sample_info,
@@ -35,12 +37,14 @@ from .updates import (
 __all__ = [
     "create_db",
     "AmpliconBarcodeRosterEntry",
+    "AmpliconSampleRow",
     "IlluminaSampleRow",
     "KatharoseqSampleInfo",
     "PacbioSampleRow",
     "PlatformSampleInfo",
     "add_katharoseq_sample",
     "get_amplicon_barcode_roster",
+    "get_amplicon_sample_info",
     "get_illumina_sample_info",
     "get_katharoseq_sample_info",
     "get_pacbio_sample_info",